Изменение характеристик противников
Здоровье и урон теперь зависят от уровня
