+++ LabaLegacy/src/main/java/Characters/Action.java
@@ -1,4 +1,4 @@
-package mortalkombatbversion.Characters;
+package Characters;
 
 public enum Action {
     /**
