+++ LabaLegacy/src/main/java/Characters/BehaviorPattern.java
@@ -1,4 +1,4 @@
-package mortalkombatbversion.Characters;
+package Characters;
 
 import java.util.List;
 import java.util.Random;
