+++ LabaLegacy/src/main/java/Characters/Debuff.java
@@ -1,4 +1,4 @@
-package mortalkombatbversion.Characters;
+package Characters;
 
 public class Debuff {
