+++ LabaLegacy/src/main/java/Characters/EnemyFabric.java
@@ -1,0 +1,13 @@
+package Characters;
+
+public class EnemyFabric {
+    public Player createEnemy(EnemyType type, int level) {
+        switch (type) {
+            case TANK:    return new Baraka(level);
+            case MAGE:    return new SubZero(level);
+            case FIGHTER: return new LiuKang(level);
+            case SOLDIER: return new SonyaBlade(level);
+            default: throw new IllegalArgumentException("Неизвестный тип врага");
+        }
+    }
+}
