+++ LabaLegacy/src/main/java/Characters/EnemyType.java
@@ -1,4 +1,4 @@
-package mortalkombatbversion.Characters;
+package Characters;
 
 public enum EnemyType {
      TANK, MAGE, FIGHTER, SOLDIER
