+++ LabaLegacy/src/main/java/Characters/LiuKang.java
@@ -1,4 +1,4 @@
-package mortalkombatbversion.Characters;
+package Characters;
 
 public class LiuKang extends Enemy {
