+++ LabaLegacy/src/main/java/Characters/ShaoKahn.java
@@ -1,4 +1,4 @@
-package mortalkombatbversion.Characters;
+package Characters;
 
 public class ShaoKahn extends Enemy {
