+++ LabaLegacy/src/main/java/Characters/SonyaBlade.java
@@ -1,4 +1,4 @@
-package mortalkombatbversion.Characters;
+package Characters;
 
 public class SonyaBlade extends Enemy {
