+++ LabaLegacy/src/main/java/mortalkombatbversion/Fight.java
@@ -1,204 +1,169 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
- */
 package mortalkombatbversion;
 
-//ADD IMAGE!!!
-import mortalkombatbversion.Characters.Player;
-import mortalkombatbversion.Characters.ShaoKahn;
-import mortalkombatbversion.Characters.EnemyFabric;
-import java.util.ArrayList;
-import javax.swing.JDialog;
-import javax.swing.JFrame;
 import javax.swing.JLabel;
-import javax.swing.JProgressBar;
-import javax.swing.JRadioButton;
+import Characters.Action;
+import Characters.Enemy;
+import Characters.Human;
+import Characters.Player;
 
-/**
- *
- * @author Мария
- */
 public class Fight {
 
-    ChangeTexts change = new ChangeTexts();
-    int kind_attack[] = {0};
-    int experiences[] = {40, 90, 180, 260, 410};
-    EnemyFabric fabric = new EnemyFabric();
-    int i = 1;
-    int k = -1;
-    int stun = 0;
-    double v = 0.0;
+    private Human player;
+    private Enemy enemy;
+    private int stunStatus = 0; // 0 - нет оглушения, 1 - игрок оглушен, 2 - враг оглушен
+    private int playerWins = 0;
+    private int currentLevel = 1;
+    private JLabel logLabel; // Для вывода сообщений
 
-    public void Move(Player p1, Player p2, JLabel l, JLabel l2) {
-        if (stun == 1) {
-            p1.setAttack(-1);
+    public Fight(Human player, Enemy enemy, JLabel logLabel) {
+        this.player = player;
+        this.enemy = enemy;
+        this.logLabel = logLabel;
+    }
+
+    public void processRound(Action playerAction) {
+
+        // Ход игрока
+        if (stunStatus == 1) {
+            logLabel.setText(player.getName() + " оглушен и пропускает ход!");
+            stunStatus = 0;
+        } else {
+            player.updateDebuff();
+            Action enemyReaction = enemy.getNextAction();
+            resolveActions(playerAction, enemyReaction, true);
         }
-        switch (Integer.toString(p1.getAttack()) + Integer.toString(p2.getAttack())) {
-            case "10":
-                v = Math.random();
-                if (p1 instanceof ShaoKahn & v < 0.15) {
-                    p2.setHealth(-(int) (p1.getDamage() * 0.5));
-                    l2.setText("Your block is broken");
 
+        //Ход компьютера
+        if (enemy.getHealth() > 0) {
+            if (stunStatus == 2) {
+                logLabel.setText(enemy.getName() + " оглушен и пропускает ход!");
+                stunStatus = 0;
+            } else {
+                enemy.updateDebuff();
+                Action enemyAction = enemy.getNextAction();
+                Action playerReaction = player.chooseAction(enemy);
+                resolveActions(enemyAction, playerReaction, false);
+            }
+        }
+
+        // Проверка победы
+        if (enemy.getHealth() <= 0) {
+            handleVictory();
+        } else if (player.getHealth() <= 0) {
+            handleDefeat();
+        }
+    }
+
+    private void resolveActions(Action initiatorAction, Action responderAction, boolean isPlayerInitiator) {
+        Player initiator;
+        Player responder;
+
+        if (isPlayerInitiator) {
+            initiator = player;
+            responder = enemy;
+        } else {
+            initiator = enemy;
+            responder = player;
+        }
+
+        switch (initiatorAction) {
+            case ATTACK:
+                if (responderAction == Action.DEFEND) {
+                    // Атака + Защита = Контрудар вполсилы
+                    int counterDamage = (int) (responder.getDamage() * 0.5);
+                    initiator.changeHealth(-counterDamage);
+                    logLabel.setText(responder.getName() + " контратаковал на " + counterDamage + " урона!");
                 } else {
-                    p1.setHealth(-(int) (p2.getDamage() * 0.5));
-                    l2.setText(p2.getName() + " counterattacked");
+                    // Атака + Атака = Инициатор наносит урон
+                    responder.changeHealth(-initiator.getDamage());
+                    logLabel.setText(initiator.getName() + " атаковал на " + initiator.getDamage() + " урона!");
                 }
                 break;
-            case "11":
-                p2.setHealth(-p1.getDamage());
-                l2.setText(p1.getName() + " attacked");
-                break;
-            case "00":
-                v = Math.random();
-                if (v <= 0.5) {
-                    stun = 1;
+
+            case DEFEND:
+                if (responderAction == Action.DEFEND) {
+                    // Защита + Защита = 50% шанс оглушения
+                    if (Math.random() < 0.5) {
+                        if (isPlayerInitiator) {
+                            stunStatus = 2; // враг оглушен
+                            logLabel.setText("Оба защищались! " + enemy.getName() + " оглушен!");
+                        } else {
+                            stunStatus = 1; // игрок оглушен
+                            logLabel.setText("Оба защищались! " + player.getName() + " оглушен!");
+                        }
+                    }
                 }
-                l2.setText("Both defended themselves");
                 break;
-            case "01":
-                l2.setText(p1.getName() + " didn't attacked");
-                break;
-            case "-10":
-                l.setText(p1.getName() + " was stunned");
-                stun = 0;
-                l2.setText(p2.getName() + " didn't attacked");
-                break;
-            case "-11":
-                p1.setHealth(-p2.getDamage());
-                l.setText(p1.getName() + " was stunned");
-                stun = 0;
-                l2.setText(p2.getName() + " attacked");
-                break;
-        }
-    }
 
-    public void Hit(Player human, Player enemy, int a, JLabel label,
-            JLabel label2, JDialog dialog, JLabel label3, CharacterAction action,
-            JProgressBar pr1, JProgressBar pr2, JDialog dialog1,
-            JDialog dialog2, JFrame frame, ArrayList<Result> results,
-            JLabel label4, JLabel label5, JLabel label6, JLabel label7,
-            JLabel label8, Items[] items, JRadioButton rb) {
-        label7.setText("");
-        human.setAttack(a);
-
-        if (k < kind_attack.length - 1) {
-            k++;
-        } else {
-            kind_attack = action.ChooseBehavior(enemy, action);
-            k = 0;
-        }
-        enemy.setAttack(kind_attack[k]);
-        if (i % 2 == 1) {
-            Move(human, enemy, label7, label8);
-        } else {
-            Move(enemy, human, label7, label8);
-        }
-        i++;
-        change.RoundTexts(human, enemy, label, label2, i, label6);
-        action.HP(human, pr1);
-        action.HP(enemy, pr2);
-        if (human.getHealth() <= 0 & items[2].getCount() > 0) {
-            human.setNewHealth((int) (human.getMaxHealth() * 0.05));
-            items[2].setCount(-1);
-            action.HP(human, pr1);
-            label2.setText(human.getHealth() + "/" + human.getMaxHealth());
-            rb.setText(items[2].getName() + ", " + items[2].getCount() + " шт");
-            label7.setText("Вы воскресли");
-        }
-        if (human.getHealth() <= 0 | enemy.getHealth() <= 0) {
-            if (((Human) human).getWin() == 11) {
-                EndFinalRound(((Human) human), action, results, dialog1, dialog2,
-                        frame, label4, label5);
-            } else {
-                EndRound(human, enemy, dialog, label3, action, items);
-            }
-        }
-    }
-
-    public void EndRound(Player human, Player enemy, JDialog dialog, JLabel label,
-            CharacterAction action, Items[] items) {
-
-        dialog.setVisible(true);
-        dialog.setBounds(300, 150, 700, 600);
-        if (human.getHealth() > 0) {
-            label.setText("You win");
-            ((Human) human).setWin();
-
-            if (enemy instanceof ShaoKahn) {
-                action.AddItems(38, 23, 8, items);
-                action.AddPointsBoss(((Human) human), action.getEnemyes());
-            } else {
-                action.AddItems(25, 15, 5, items);
-                action.AddPoints(((Human) human), action.getEnemyes());
-            }
-        } else {
-            label.setText(enemy.getName() + " win");
-        }
-
-        i = 1;
-        k = -1;
-        kind_attack = ResetAttack();
-
-    }
-
-    public void EndFinalRound(Human human, CharacterAction action,
-            ArrayList<Result> results, JDialog dialog1, JDialog dialog2, JFrame frame,
-            JLabel label1, JLabel label2) {
-        String text = "Победа не на вашей стороне";
-        if (human.getHealth() > 0) {
-            human.setWin();
-            action.AddPoints(human, action.getEnemyes());
-            text = "Победа на вашей стороне";
-        }
-        boolean top = false;
-        if (results == null) {
-            top = true;
-        } else {
-            int i = 0;
-            for (int j = 0; j < results.size(); j++) {
-                if (human.getPoints() < results.get(j).getPoints()) {
-                    i++;
+            case WEAKEN:
+                initiator.applyWeaken(responder, responderAction);
+                if (responderAction == Action.DEFEND && responder.isWeakened()) {
+                    logLabel.setText(initiator.getName() + " ослабил " + responder.getName() + "!");
+                } else if (responderAction == Action.ATTACK) {
+                    logLabel.setText(responder.getName() + " сорвал ослабление!");
+                } else {
+                    logLabel.setText(initiator.getName() + " не может использовать ослабление!");
                 }
-            }
-            if (i < 10) {
-                top = true;
-            }
+                break;
         }
-        if (top) {
-            dialog1.setVisible(true);
-            dialog1.setBounds(150, 150, 600, 500);
-            label1.setText(text);
-        } else {
-            dialog2.setVisible(true);
-            dialog2.setBounds(150, 150, 470, 360);
-            label2.setText(text);
-        }
-        frame.dispose();
     }
 
-    public int[] ResetAttack() {
-        int a[] = {0};
-        return a;
-    }
+    private void handleVictory() {
+        playerWins++;
+        int expGain = calculateExperience();
+        player.addExperience(expGain);
+        player.addPoints(calculatePoints());
 
-    public Player NewRound(Player human, JLabel label, JProgressBar pr1,
-            JProgressBar pr2, JLabel label2, JLabel text, JLabel label3, CharacterAction action) {
+        logLabel.setText(player.getName() + " побеждает! +" + expGain + " опыта");
 
-        Player enemy1 = null;
-        if (((Human) human).getWin() == 6 | ((Human) human).getWin() == 11) {
-            enemy1 = action.ChooseBoss(label, label2, text, label3, human.getLevel());
-        } else {
-            enemy1 = action.ChooseEnemy(label, label2, text, label3);
+        // Проверка перехода на новый уровень
+        if (playerWins >= getWinsForLevel(currentLevel)) {
+            currentLevel++;
+            player.levelUp();
+            logLabel.setText("Уровень повышен! Теперь уровень " + currentLevel);
         }
-        pr1.setMaximum(human.getMaxHealth());
-        pr2.setMaximum(enemy1.getMaxHealth());
-        human.setNewHealth(human.getMaxHealth());
-        enemy1.setNewHealth(enemy1.getMaxHealth());
-        action.HP(human, pr1);
-        action.HP(enemy1, pr2);
-        return enemy1;
     }
 
+    private void handleDefeat() {
+        logLabel.setText("Вы проиграли! Игра окончена.");
+    }
+
+    private int calculateExperience() {
+        // Опыт зависит от уровня врага и полученного урона
+        return 10 + currentLevel * 5;
+    }
+
+    private int calculatePoints() {
+        // Очки зависят от оставшегося здоровья
+        return (int) (player.getHealth() / (double) player.getMaxHealth() * 100);
+    }
+
+    private int getWinsForLevel(int level) {
+        return switch (level) {
+            case 1 ->
+                2;
+            case 2 ->
+                4;
+            case 3 ->
+                7;
+            case 4 ->
+                9;
+            case 5 ->
+                12;
+            default ->
+                15;
+        };
+    }
+
+    public int getCurrentLevel() {
+        return currentLevel;
+    }
+
+    public int getPlayerWins() {
+        return playerWins;
+    }
+
+    public int getWinsForNextLevel() {
+        return getWinsForLevel(currentLevel);
+    }
 }
