+++ LabaLegacy/src/main/java/mortalkombatbversion/mainIdea.java
@@ -1,0 +1,15 @@
+package mortalkombatbversion;
+
+import Characters.EnemyFabric;
+import Characters.Human;
+
+public class mainIdea {
+    private int locationQuantity;
+    private Human player = new Human();
+    private EnemyFabric enemyFabric = new EnemyFabric();
+    
+    public mainIdea(int quantityLocation){
+        this.locationQuantity = quantityLocation;
+    }
+    
+}
